+++ CHANGELOG.rst
@@ -24,7 +24,7 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 Added
 `````
 
-...
+- ``libstored.protocol.ArqLayer`` for general-purpose ARQ.
 
 Fixed
 `````